Fuzz the server logic.
A single PROTMESSID_NETW_TRANSPORT_PROPS message is allowed as a prefix
to another message, in order to allow the fuzzer to get into the audio
processing code. (This seems to work best if you seed it with a valid
PROPS + audio message pair.)

The same could be done for the client logic, but it'd need the audio
device code stubbing out.